Addressing review comments from Andrej
